refactor(directives): consolidate directives in /directives
